requirements.txt
SQLAlchemy==1.3.23
pandas==1.2.2
python-dotenv==0.15.0
pymysql==1.0.2